feat: address PR's comments
