+++ Cargo.toml
@@ -7,6 +7,7 @@ description = "Persistent Merkle Tree in Rust"
 
 [dependencies]
 rayon = { version = "1.12.0", optional = true }
+parking_lot = "0.12.5"
 thiserror = "2.0.18"
 
 [dev-dependencies]
